Tried to fix command statements text field editing
the problem is, that the drag drop of the item avoids editing the text
field and clicking inside the textfield. I disabled the drag and drop
when the item is inside the code area.
Next steps: add button for delete commands manually and perhaps button
to enable drag drop again to change the order of commands inside the
code area.